as int

+++ CT_CLIP/ct_clip/utils.py
@@ -74,7 +74,7 @@ def vis_3d_img_list(img_list, slice_ratio_list=[0.25, 0.5, 0.75], img_name="imag
         # save to the wandb folder
         os.makedirs("wandb", exist_ok=True)
         save_path = f"wandb/{image_name}.png"
-        cv2.imwrite(save_path, slices * 255)
+        cv2.imwrite(save_path, (slices * 255).astype(np.uint8))
 
         # Convert the figure to a WandB image object
         wandb_img = wandb.Image(slices, caption=image_name)
